rename in pretty-format/src/plugins/lib/escape_html.js
