+++ packages/pretty-format/src/plugins/lib/escapeHTML.js
